+++ cypress/e2e/demo/errors.cy.ts
@@ -1,8 +1,8 @@
 context('Errors', () => {
   function checkErrorReported(type: string, value: string, expectStacktrace = true) {
-    cy.waitExceptions((events) => {
-      expect(events).to.have.lengthOf(1);
-      const exception = events[0];
+    cy.waitExceptions((evts) => {
+      expect(evts).to.have.lengthOf(1);
+      const exception = evts[0];
       expect(exception).property('type').to.equal(type);
       expect(exception).property('value').to.equal(value);
       if (expectStacktrace) {
