+++ cypress/e2e/demo/measurements.cy.ts
@@ -5,9 +5,9 @@ context('Measurements', () => {
     // load blank page to force page unload and sending
     cy.loadBlank();
 
-    cy.waitMeasurements((events) => {
-      expect(events).to.have.lengthOf(4);
-      const vitals = events.flatMap((evt) => Object.keys(evt.values));
+    cy.waitMeasurements((evts) => {
+      expect(evts).to.have.lengthOf(4);
+      const vitals = evts.flatMap((evt) => Object.keys(evt.values));
       EXPECTED_VITALS.forEach((vital) => expect(vitals).to.include(vital));
     }, 4);
   });
