+++ packages/core/src/testUtils/mockConfig.ts
@@ -1,19 +1,22 @@
 import type { Config } from '../config';
+import { defaultInternalLoggerLevel } from '../internalLogger';
 import { mockStacktraceParser } from './mockStacktraceParser';
 
 export function mockConfig(overrides: Partial<Config> = {}): Config {
   return {
-    globalObjectKey: 'grafanaAgent',
-    instrumentations: [],
-    preventGlobalExposure: true,
-    transports: [],
-    metas: [],
     app: {
       name: 'test',
       version: '1.0.0',
     },
+    globalObjectKey: 'grafanaAgent',
+    internalLoggerLevel: defaultInternalLoggerLevel,
+    instrumentations: [],
+    metas: [],
     parseStacktrace: mockStacktraceParser,
     paused: false,
+    preventGlobalExposure: true,
+    transports: [],
+    unpatchedConsole: console,
     ...overrides,
   };
 }
