+++ packages/tracing-web/src/agentExporter.ts
@@ -1,13 +1,11 @@
-import type { Agent, TraceEvent } from '@grafana/agent-core';
+import type { TraceEvent } from '@grafana/agent-core';
 import { ExportResultCode } from '@opentelemetry/core';
 import type { ExportResult } from '@opentelemetry/core';
 import { createExportTraceServiceRequest } from '@opentelemetry/otlp-transformer';
 import type { IExportTraceServiceRequest } from '@opentelemetry/otlp-transformer';
 import type { ReadableSpan, SpanExporter } from '@opentelemetry/sdk-trace-base';
 
-interface GrafanaAgentTraceExporterConfig {
-  agent: Agent;
-}
+import type { GrafanaAgentTraceExporterConfig } from './types';
 
 export class GrafanaAgentTraceExporter implements SpanExporter {
   constructor(private config: GrafanaAgentTraceExporterConfig) {}
