+++ packages/tracing-web/src/index.ts
@@ -1,4 +1,7 @@
 export { GrafanaAgentTraceExporter } from './agentExporter';
 
-export { TracingInstrumentation, getDefaultOTELInstrumentations } from './instrumentation';
-export type { TracingInstrumentationOptions } from './instrumentation';
+export { getDefaultOTELInstrumentations } from './getDefaultOTELInstrumentations';
+
+export { TracingInstrumentation } from './instrumentation';
+
+export type { GrafanaAgentTraceExporterConfig, TracingInstrumentationOptions } from './types';
