+++ packages/web/src/config/index.ts
@@ -1,0 +1,5 @@
+export { getWebInstrumentations } from './getWebInstrumentations';
+
+export { makeCoreConfig } from './makeCoreConfig';
+
+export type { BrowserConfig, GetWebInstrumentationsOptions } from './types';
