+++ packages/web/src/config/types.ts
@@ -1,0 +1,10 @@
+import type { Config } from '@grafana/agent-core';
+
+export interface BrowserConfig extends Partial<Omit<Config, 'app' | 'parseStacktrace'>>, Pick<Config, 'app'> {
+  url?: string;
+  apiKey?: string;
+}
+
+export interface GetWebInstrumentationsOptions {
+  captureConsole?: boolean;
+}
