+++ packages/web/src/instrumentations/console/instrumentation.ts
@@ -1,0 +1,33 @@
+import { allLogLevels, BaseInstrumentation, LogLevel, VERSION } from '@grafana/agent-core';
+
+import type { ConsoleInstrumentationOptions } from './types';
+
+export class ConsoleInstrumentation extends BaseInstrumentation {
+  readonly name = '@grafana/agent-web:instrumentation-console';
+  readonly version = VERSION;
+
+  static defaultDisabledLevels: LogLevel[] = [LogLevel.DEBUG, LogLevel.TRACE, LogLevel.LOG];
+
+  constructor(private options: ConsoleInstrumentationOptions = {}) {
+    super();
+  }
+
+  initialize() {
+    this.logDebug('Initializing...\n', this.options);
+
+    allLogLevels
+      .filter((level) => !(this.options.disabledLevels ?? ConsoleInstrumentation.defaultDisabledLevels).includes(level))
+      .forEach((level) => {
+        /* eslint-disable-next-line no-console */
+        console[level] = (...args) => {
+          try {
+            this.agent.api.pushLog(args, { level });
+          } catch (err) {
+            this.logError(err);
+          } finally {
+            this.agent.unpatchedConsole[level](...args);
+          }
+        };
+      });
+  }
+}
