+++ packages/web/src/instrumentations/console/types.ts
@@ -1,0 +1,5 @@
+import type { LogLevel } from '@grafana/agent-core';
+
+export interface ConsoleInstrumentationOptions {
+  disabledLevels?: LogLevel[];
+}
