--- packages/web/src/instrumentations/errors/extendedError.ts
@@ -1,5 +1,0 @@
-export interface ExtendedError extends Error {
-  columnNumber?: number;
-  framesToPop?: number;
-  stacktrace?: Error['stack'];
-}
