+++ packages/web/src/instrumentations/errors/stackFrames/buildStackFrame.ts
@@ -1,0 +1,25 @@
+import type { ExceptionStackFrame } from '@grafana/agent-core';
+
+import { unknownString } from './const';
+
+export function buildStackFrame(
+  filename: string | undefined,
+  func: string | undefined,
+  lineno: number | undefined,
+  colno: number | undefined
+): ExceptionStackFrame {
+  const stackFrame: ExceptionStackFrame = {
+    filename: filename || document.location.href,
+    function: func || unknownString,
+  };
+
+  if (lineno !== undefined) {
+    stackFrame.lineno = lineno;
+  }
+
+  if (colno !== undefined) {
+    stackFrame.colno = colno;
+  }
+
+  return stackFrame;
+}
