+++ packages/web/src/instrumentations/errors/stackFrames/index.ts
@@ -1,0 +1,7 @@
+export { buildStackFrame } from './buildStackFrame';
+
+export { getDataFromSafariExtensions } from './getDataFromSafariExtensions';
+
+export { getStackFramesFromError } from './getStackFramesFromError';
+
+export { parseStacktrace } from './parseStacktrace';
