+++ packages/web/src/instrumentations/errors/stackFrames/parseStacktrace.ts
@@ -1,0 +1,9 @@
+import type { ExtendedError, Stacktrace } from '@grafana/agent-core';
+
+import { getStackFramesFromError } from './getStackFramesFromError';
+
+export function parseStacktrace(error: ExtendedError): Stacktrace {
+  return {
+    frames: getStackFramesFromError(error),
+  };
+}
