+++ packages/web/src/instrumentations/index.ts
@@ -1,3 +1,13 @@
-export * from './errors';
-export * from './web-vitals';
-export * from './console';
+export { ConsoleInstrumentation } from './console';
+export type { ConsoleInstrumentationOptions } from './console';
+
+export {
+  buildStackFrame,
+  ErrorsInstrumentation,
+  getDataFromSafariExtensions,
+  getStackFramesFromError,
+  parseStacktrace,
+} from './errors';
+export type { ErrorEvent, ExtendedPromiseRejectionEvent } from './errors';
+
+export { WebVitalsInstrumentation } from './webVitals';
