+++ packages/web/src/instrumentations/webVitals/index.ts
@@ -1,0 +1,1 @@
+export { WebVitalsInstrumentation } from './instrumentation';
