+++ packages/web/src/metas/index.ts
@@ -1,2 +1,5 @@
-export * from './browser';
-export * from './page';
+export { browserMeta } from './browser';
+
+export { defaultMetas } from './const';
+
+export { pageMeta } from './page';
