+++ packages/web/src/metas/page/index.ts
@@ -1,7 +1,1 @@
-import type { Meta, MetaItem } from '@grafana/agent-core';
-
-export const pageMeta: MetaItem<Pick<Meta, 'page'>> = () => ({
-  page: {
-    url: location.href,
-  },
-});
+export { pageMeta } from './meta';
