+++ packages/web/src/metas/page/meta.ts
@@ -1,0 +1,7 @@
+import type { Meta, MetaItem } from '@grafana/agent-core';
+
+export const pageMeta: MetaItem<Pick<Meta, 'page'>> = () => ({
+  page: {
+    url: location.href,
+  },
+});
