+++ packages/web/src/transports/console/types.ts
@@ -1,0 +1,5 @@
+import type { LogLevel } from '@grafana/agent-core';
+
+export interface ConsoleTransportOptions {
+  level?: LogLevel;
+}
