+++ packages/web/src/transports/index.ts
@@ -1,2 +1,5 @@
-export * from './console';
-export * from './fetch';
+export { ConsoleTransport } from './console';
+export type { ConsoleTransportOptions } from './console';
+
+export { FetchTransport } from './fetch';
+export type { FetchTransportOptions, FetchTransportRequestOptions } from './fetch';
